Crate solana_account

The Solana Account type.

Structs

Account

An Account with data that is stored on chain

AccountSharedData

An Account with data that is stored on chain This will be the in-memory representation of the ‘Account’ struct data. The existing ‘Account’ structure cannot easily change due to downstream projects.

Constants

DUMMY_INHERITABLE_ACCOUNT_FIELDS

PROGRAM_OWNERS

Replacement for the executable flag: An account being owned by one of these contains a program.

Traits

ReadableAccount

WritableAccount

Functions

accounts_equal

Compares two ReadableAccounts

create_is_signer_account_infos

Create AccountInfos

Type Aliases

InheritableAccountFields